requirements.txt
# === Core ML stack ===
torch>=2.4.0
transformers>=4.46.0
accelerate>=1.0.0
bitsandbytes>=0.44.0


numpy>=1.26.0
scipy>=1.13.0
pandas>=2.2.0
scikit-learn>=1.5.0

pymupdf>=1.24.0

pydantic>=2.9.0
pyyaml>=6.0.2
python-dotenv>=1.0.1


tqdm>=4.66.0


matplotlib>=3.9.0
seaborn>=0.13.0


pytest>=8.3.0